+++ BERT/code/extract_features.py
@@ -12,7 +12,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Extract pre-computed feature vectors from BERT."""
+"""Extract pre-computed feature vectors from BERT_checkpoint."""
 
 from __future__ import absolute_import
 from __future__ import division
@@ -23,8 +23,7 @@ import collections
 import json
 import re
 
-import modeling
-import tokenization
+from BERT import modeling, tokenization
 import tensorflow as tf
 
 flags = tf.flags
@@ -39,7 +38,7 @@ flags.DEFINE_string("layers", "-1,-2,-3,-4", "")
 
 flags.DEFINE_string(
     "bert_config_file", None,
-    "The config json file corresponding to the pre-trained BERT model. "
+    "The config json file corresponding to the pre-trained BERT_checkpoint model. "
     "This specifies the model architecture.")
 
 flags.DEFINE_integer(
@@ -50,10 +49,10 @@ flags.DEFINE_integer(
 
 flags.DEFINE_string(
     "init_checkpoint", None,
-    "Initial checkpoint (usually from a pre-trained BERT model).")
+    "Initial checkpoint (usually from a pre-trained BERT_checkpoint model).")
 
 flags.DEFINE_string("vocab_file", None,
-                    "The vocabulary file that the BERT model was trained on.")
+                    "The vocabulary file that the BERT_checkpoint model was trained on.")
 
 flags.DEFINE_bool(
     "do_lower_case", True,
@@ -228,7 +227,7 @@ def convert_examples_to_features(examples, seq_length, tokenizer):
       if len(tokens_a) > seq_length - 2:
         tokens_a = tokens_a[0:(seq_length - 2)]
 
-    # The convention in BERT is:
+    # The convention in BERT_checkpoint is:
     # (a) For sequence pairs:
     #  tokens:   [CLS] is this jack ##son ##ville ? [SEP] no it is not . [SEP]
     #  type_ids: 0     0  0    0    0     0       0 0     1  1  1  1   1 1
