+++ BERT/code/optimization.py
@@ -22,6 +22,7 @@ import re
 import tensorflow as tf
 
 
+
 def create_optimizer(loss, init_lr, num_train_steps, num_warmup_steps, use_tpu):
   """Creates an optimizer training op."""
   global_step = tf.train.get_or_create_global_step()
