+++ BERT/code/run_pretraining.py
@@ -12,15 +12,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Run masked LM/next sentence masked_lm pre-training for BERT."""
+"""Run masked LM/next sentence masked_lm pre-training for BERT_checkpoint."""
 
 from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
 import os
-import modeling
-import optimization
+from BERT import modeling, optimization
 import tensorflow as tf
 
 flags = tf.flags
@@ -30,7 +29,7 @@ FLAGS = flags.FLAGS
 ## Required parameters
 flags.DEFINE_string(
     "bert_config_file", None,
-    "The config json file corresponding to the pre-trained BERT model. "
+    "The config json file corresponding to the pre-trained BERT_checkpoint model. "
     "This specifies the model architecture.")
 
 flags.DEFINE_string(
@@ -44,7 +43,7 @@ flags.DEFINE_string(
 ## Other parameters
 flags.DEFINE_string(
     "init_checkpoint", None,
-    "Initial checkpoint (usually from a pre-trained BERT model).")
+    "Initial checkpoint (usually from a pre-trained BERT_checkpoint model).")
 
 flags.DEFINE_integer(
     "max_seq_length", 128,
