+++ onSTILTs/configFiles_STILTs/InsufficientArg.sh
@@ -1,0 +1,11 @@
+#!/bin/sh
+
+# Einfaches Beispiel
+echo Hallo, Welt!
+echo $PATH
+echo Datum, Uhrzeit und Arbeitsverzeichnis:
+date
+pwd
+echo Uebergabe-Parameter: $*
+
+cmd /K
